src/Python/requirements.txt
prophet
mysql-connector-python
pandas
numpy
matplotlib
scikit-learn
pyyaml
